+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- SSE support: SSEClient/SSEConnection with parsing, reconnection and tests.
+
 ## [0.26.0] 2025-09-26
 
 ### Added